+++ src/main/java/org/apache/aurora/scheduler/discovery/ServiceDiscoveryModule.java
@@ -13,24 +13,29 @@
  */
 package org.apache.aurora.scheduler.discovery;
 
+import java.io.File;
+import java.io.IOException;
+import java.net.InetSocketAddress;
 import java.util.List;
 
 import javax.inject.Singleton;
 
-import com.google.common.base.Optional;
-import com.google.inject.Exposed;
-import com.google.inject.PrivateModule;
+import com.google.common.base.Throwables;
+import com.google.common.collect.ImmutableList;
+import com.google.common.io.Files;
+import com.google.common.util.concurrent.AbstractIdleService;
+import com.google.inject.AbstractModule;
+import com.google.inject.Inject;
+import com.google.inject.Module;
+import com.google.inject.Provider;
 import com.google.inject.Provides;
+import com.google.inject.binder.LinkedBindingBuilder;
 
-import org.apache.aurora.common.net.pool.DynamicHostSet;
-import org.apache.aurora.common.thrift.ServiceInstance;
-import org.apache.aurora.common.zookeeper.Credentials;
-import org.apache.aurora.common.zookeeper.ServerSetImpl;
-import org.apache.aurora.common.zookeeper.SingletonService;
-import org.apache.aurora.common.zookeeper.SingletonServiceImpl;
-import org.apache.aurora.common.zookeeper.ZooKeeperClient;
+import org.apache.aurora.common.application.ShutdownRegistry;
+import org.apache.aurora.common.base.MorePreconditions;
 import org.apache.aurora.common.zookeeper.ZooKeeperUtils;
-import org.apache.aurora.scheduler.app.ServiceGroupMonitor;
+import org.apache.aurora.common.zookeeper.testing.ZooKeeperTestServer;
+import org.apache.aurora.scheduler.SchedulerServicesModule;
 import org.apache.zookeeper.data.ACL;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -38,32 +43,62 @@ import org.slf4j.LoggerFactory;
 import static java.util.Objects.requireNonNull;
 
 /**
- * Binding module for utilities to advertise the network presence of the scheduler.
+ * Creates a Guice module that binds leader election and leader discovery components.
  */
-public class ServiceDiscoveryModule extends PrivateModule {
+public class ServiceDiscoveryModule extends AbstractModule {
 
-  private static final Logger LOG = LoggerFactory.getLogger(ServiceDiscoveryModule.class);
+  private static final Logger LOG = LoggerFactory.getLogger(CommonsServiceDiscoveryModule.class);
 
-  private final String serverSetPath;
-  private final Optional<Credentials> zkCredentials;
+  private final ZooKeeperConfig zooKeeperConfig;
+  private final String discoveryPath;
 
-  public ServiceDiscoveryModule(String serverSetPath, Optional<Credentials> zkCredentials) {
-    this.serverSetPath = requireNonNull(serverSetPath);
-    this.zkCredentials = requireNonNull(zkCredentials);
+  /**
+   * Creates a Guice module that will bind a
+   * {@link org.apache.aurora.common.zookeeper.SingletonService} for scheduler leader election and a
+   * {@link org.apache.aurora.scheduler.app.ServiceGroupMonitor} that can be used to find the
+   * leading scheduler.
+   *
+   * @param zooKeeperConfig The ZooKeeper client configuration to use to interact with ZooKeeper.
+   * @param discoveryPath The ZooKeeper path to use to host leader election and service discovery
+   *                      nodes under.
+   */
+  public ServiceDiscoveryModule(ZooKeeperConfig zooKeeperConfig, String discoveryPath) {
+    this.zooKeeperConfig = requireNonNull(zooKeeperConfig);
+    this.discoveryPath = MorePreconditions.checkNotBlank(discoveryPath);
   }
 
   @Override
   protected void configure() {
-    requireBinding(ZooKeeperClient.class);
+    LinkedBindingBuilder<Iterable<InetSocketAddress>> clusterBinder =
+        bind(ServiceDiscoveryBindings.ZOO_KEEPER_CLUSTER_KEY);
 
-    bind(ServiceGroupMonitor.class).to(CommonsServerGroupMonitor.class).in(Singleton.class);
-    expose(ServiceGroupMonitor.class);
+    if (zooKeeperConfig.isInProcess()) {
+      requireBinding(ShutdownRegistry.class);
+      File tempDir = Files.createTempDir();
+      bind(ZooKeeperTestServer.class).toInstance(new ZooKeeperTestServer(tempDir, tempDir));
+      SchedulerServicesModule.addAppStartupServiceBinding(binder()).to(TestServerService.class);
+
+      clusterBinder.toProvider(LocalZooKeeperClusterProvider.class);
+    } else {
+      clusterBinder.toInstance(zooKeeperConfig.getServers());
+    }
+
+    install(discoveryModule());
+  }
+
+  private Module discoveryModule() {
+    if (zooKeeperConfig.isUseCurator()) {
+      return new CuratorServiceDiscoveryModule(discoveryPath, zooKeeperConfig);
+    } else {
+      return new CommonsServiceDiscoveryModule(discoveryPath, zooKeeperConfig);
+    }
   }
 
   @Provides
   @Singleton
+  @ServiceDiscoveryBindings.ZooKeeper
   List<ACL> provideAcls() {
-    if (zkCredentials.isPresent()) {
+    if (zooKeeperConfig.getCredentials().isPresent()) {
       return ZooKeeperUtils.EVERYONE_READ_CREATOR_ALL;
     } else {
       LOG.warn("Running without ZooKeeper digest credentials. ZooKeeper ACLs are disabled.");
@@ -71,30 +106,49 @@ public class ServiceDiscoveryModule extends PrivateModule {
     }
   }
 
-  @Provides
-  @Singleton
-  ServerSetImpl provideServerSet(ZooKeeperClient client, List<ACL> zooKeeperAcls) {
-    return new ServerSetImpl(client, zooKeeperAcls, serverSetPath);
+  /**
+   * A service to wrap ZooKeeperTestServer.  ZooKeeperTestServer is not a service itself because
+   * some tests depend on stop/start routines that do not no-op, like startAsync and stopAsync may.
+   */
+  private static class TestServerService extends AbstractIdleService {
+    private final ZooKeeperTestServer testServer;
+
+    @Inject
+    TestServerService(ZooKeeperTestServer testServer) {
+      this.testServer = requireNonNull(testServer);
+    }
+
+    @Override
+    protected void startUp() {
+      // We actually start the test server on-demand rather than with the normal lifecycle.
+      // This is because a ZooKeeperClient binding is needed before scheduler services are started.
+    }
+
+    @Override
+    protected void shutDown() {
+      testServer.stop();
+    }
   }
 
-  @Provides
-  @Singleton
-  DynamicHostSet<ServiceInstance> provideServerSet(ServerSetImpl serverSet) {
-    // Used for a type re-binding of the server set.
-    return serverSet;
-  }
+  private static class LocalZooKeeperClusterProvider
+      implements Provider<Iterable<InetSocketAddress>> {
 
-  // NB: We only take a ServerSetImpl instead of a ServerSet here to simplify binding.
-  @Provides
-  @Singleton
-  @Exposed
-  SingletonService provideSingletonService(
-      ZooKeeperClient client,
-      ServerSetImpl serverSet,
-      List<ACL> zookeeperAcls) {
+    private final ZooKeeperTestServer testServer;
 
-    return new SingletonServiceImpl(
-        serverSet,
-        SingletonServiceImpl.createSingletonCandidate(client, serverSetPath, zookeeperAcls));
+    @Inject
+    LocalZooKeeperClusterProvider(ZooKeeperTestServer testServer) {
+      this.testServer = requireNonNull(testServer);
+    }
+
+    @Override
+    public Iterable<InetSocketAddress> get() {
+      try {
+        testServer.startNetwork();
+      } catch (IOException | InterruptedException e) {
+        throw Throwables.propagate(e);
+      }
+      return ImmutableList.of(
+          InetSocketAddress.createUnresolved("localhost", testServer.getPort()));
+    }
   }
 }